+++ examples/data-sources/iosxr_prefix_set/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_prefix_set" "example" {
+  set_name = "PREFIX_SET_1"
+}